added folder structures and extracted the jwt function
